refactored the project
