+++ projects/angular-material-extensions/fab-menu/src/lib/animations/menu.animation.ts
