--- projects/angular-material-extensions/fab-menu/src/lib/mat-fab-menu.component.scss
@@ -1,31 +1,0 @@
-:host {
-  display: block;
-}
-
-.fullSVG {
-  img {
-    width: 100%;
-  }
-
-  ::ng-deep span.mat-button-wrapper {
-    padding: 0;
-  }
-}
-
-.container {
-  //position: fixed;
-  //bottom: 15px;
-  //right: 15px;
-  //z-index: 100;
-
-  > div {
-    //display: flex;
-    //flex-direction: column-reverse;
-    //align-items: center;
-    margin-bottom: 5px;
-
-    button {
-      margin-bottom: 16px;
-    }
-  }
-}
