+++ projects/angular-material-extensions/fab-menu/src/lib/mat-fab-menu/mat-fab-menu.component.scss
@@ -1,0 +1,13 @@
+:host {
+  display: block;
+}
+
+.fullSVG {
+  img {
+    width: 100%;
+  }
+
+  ::ng-deep span.mat-button-wrapper {
+    padding: 0;
+  }
+}
