+++ projects/angular-material-extensions/fab-menu/src/public-api.ts
@@ -2,5 +2,7 @@
  * Public API Surface of fab-menu
  */
 
-export * from './lib/mat-fab-menu.component';
+export * from './lib/mat-fab-menu/mat-fab-menu.component';
+export * from './lib/mat-mini-fab-menu/mat-mini-fab-menu.component';
+export * from './lib/models/mat-fab-menu.model';
 export * from './lib/mat-fab-menu.module';
